rfr(rdx): rename StoreConnector, break into files
I think a better name for StoreConnector is StateStreamBuilder as it's
a widget that (potentially) rebuilds every time the stream-of-state
-changes emits an event.

I broke out the widgets into separate files for readability.
